+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 2.0.0 (upcoming)
 
-- Remove hide on screen helper
+- Added link field
+- Removed hide on screen helper
 
 ## 1.0.0 (released 2017-05-30)